requirements.txt
opencv-python~=4.10.0.84
pillow~=11.0.0
ultralytics~=8.3.55
pyqt6~=6.8.0